+++ frontend/app/components/modules/HorizontalCard/HorizontalCard.tsx
@@ -1,0 +1,17 @@
+import styles from './HorizontalCard.module.scss'
+import { ReactNode } from 'react'
+
+type Props = {
+  children: ReactNode
+  className?: string
+  title?: string
+}
+
+export const HorizontalCard = ({ children, className, title }: Props) => {
+  return (
+    <div className={styles.horizontalCard}>
+      {title && <h2 className={styles.title}>{title}</h2>}
+      <div className={`${styles.card} ${className}`}>{children}</div>
+    </div>
+  )
+}