+++ frontend/app/components/modules/ResponsiveImage/ResponsiveImage.tsx
@@ -1,0 +1,17 @@
+import Image from 'next/image'
+import styles from './ResponsiveImage.module.scss'
+
+type Props = {
+  className?: string
+  src: string
+  alt: string
+  objectFit: NonNullable<JSX.IntrinsicElements['img']['style']>['objectFit']
+}
+
+export const ResponsiveImage = ({ className, src, alt, objectFit }: Props) => {
+  return (
+    <div className={`${styles.responsiveImage} ${className}`}>
+      <Image src={src} alt={alt} layout="fill" objectFit={objectFit} />
+    </div>
+  )
+}